Add av dep and additional cb to useLocalAudioInputActivityPreview
