+++ src/index.ts
@@ -123,6 +123,7 @@ export { useSelectVideoInputDevice } from './hooks/sdk/useSelectVideoInputDevice
 export { useActiveSpeakersState } from './hooks/sdk/useActiveSpeakersState';
 export { useToggleLocalMute } from './hooks/sdk/useToggleLocalMute';
 export { useMeetingStatus } from './hooks/sdk/useMeetingStatus';
+export { useLocalAudioInputActivity } from './hooks/sdk/useLocalAudioInputActivity';
 export { useLocalAudioInputActivityPreview } from './hooks/sdk/useLocalAudioInputActivityPreview';
 export { useBandwidthMetrics } from './hooks/sdk/useBandwidthMetrics';
